Upload HarvestedLithologytoOracle-SSCNGIS.ipynb
HarvestedLithologytoOracle-SSCNGIS.ipynb was written by Mike Barnes. He documented it with these comments:

After recognizing a significant limitation with our current systems in terms of lithology capture and reuse (ie our logging form did not play well with Oracle at all), I revised the logging sheet to it more closely linked to the Oracle lookup tables. 

The latest version of the blank template is found here: 

\\prod.lan\active\proj\futurex\Common\Working\Mike\GitHub\mjb_work_code\BoreholeLithologyLoading\GenericSedLogging\Templates\LithLogTemplate-20190529.xlsx  

The code to convert it to Oracle bulk loader format is here: 

\\prod.lan\active\proj\futurex\Common\Working\Mike\GitHub\mjb_work_code\BoreholeLithologyLoading\GenericSedLogging\HarvestedLithologytoOracle-SSCNGIS.ipynb  

If making changes, keep in mind that this is a very nuanced bit of work, with lots of lookups etc. 

It's a compromise between having a logging sheet that a geologist would actually want to use vs having a sheet that is directly populating Oracle. 

The end vision was to have a sedimentary rock and sediment logging template sheet that ties in nicely to the Oracle lookups that could be easily attached to variety of spatial entities (eg fieldsites, boreholes, measured sections). The lithology description fields wouldn't vary, but the top depth, base depth and borehole name could be swapped out for a XYZ from a GPS with a fieldsite name. 

That was all on my to do list. Hopefully someone else will pick it up because ultimately we should be standard across GA for describing rocks.